schedule entries

+++ app/Http/Controllers/ScheduleEntryController.php
@@ -55,9 +55,8 @@ class ScheduleEntryController extends Controller
     public function store(Request $request)
     {
        
-
        
-      
+     
           
          $schedule = new Schedule();
      $data = new Collection($request);
@@ -87,8 +86,11 @@ class ScheduleEntryController extends Controller
             if($class_roomData['class_room']==null){
                 $class_rooms[] ="undecided";
             }
+            else{
+                $class_rooms[] = $class_roomData['class_room']; // Access the nested ID
 
-            $class_rooms[] = $class_roomData['class_room']; // Access the nested ID
+            }
+
         }
         foreach ($data['day'] as $dayData) {
             $days[] = $dayData['day']; // Access the nested ID
@@ -111,28 +113,13 @@ class ScheduleEntryController extends Controller
         $scheduleEntry->day = $days[$x];
         $scheduleEntry->start_time = $start_times[$x];
         
-        if($scheduleEntry->save())
-        $saved_entries_counter++;
-                                            }
-      
-         if($saved_entries_counter!=$numberofentries){
-           $deleted=0;
-            $scheduleEntries_toBe_deleted= ScheduleEntry::where('schedule_id',$schedule->id)->get();
-            foreach($scheduleEntries_toBe_deleted as $entrie){
-               if($entrie->delete())
-               $deleted++;
-
-                
-            }
-             //return "recieved should be :".$numberofentries."</br>inserted:".$saved_entries_counter."</br>deleted".$deleted;
-             if($deleted == $saved_entries_counter) {
-                $schedule->delete();
-
-
-                                                    }
-              
+          $scheduleEntry->save();
          
-             }
+                                            }
+            
+             //return "recieved should be :".$numberofentries."</br>inserted:".$saved_entries_counter."</br>deleted".$deleted;
+            return redirect()->route("schedules");
+             
 
              
       
@@ -159,13 +146,16 @@ class ScheduleEntryController extends Controller
         
         
     }
-
+  
+     
 
     public function edit(Schedule $schedule ){
         $subjects = Subject::get();
         $instructors = Instructor::get();
+  
 
         $scheduleEntries = ScheduleEntry::where('schedule_id',$schedule->id)->get();
+        $scheduleEntries_for_update = $scheduleEntries;
 
         $i = 0;
 
@@ -179,10 +169,124 @@ class ScheduleEntryController extends Controller
 
 
     }
+    
  
-    public function update(Request $request){
+    public function add_new( Request $request ,Schedule $schedule){
 
-     return $request;
+   
+        $data = new Collection($request);  
+       
+        $instructorIds = [];
+        $subjectIds=[];
+        $days=[];
+        $start_times=[];
+        $class_rooms=[];
+        $numberofentries =0;
+
+        foreach($data["instructor_id"] as $inst){
+             
+            $instructorIds[]= $inst["instructor_id"];
+            $numberofentries++;
+       
+    } 
+    foreach ($data['subject_id'] as $subjectData) {
+        $subjectIds[] = $subjectData['subject_id'];
+       
+    } 
+    
+    foreach ($data['class_room'] as $class_roomData) {
+        if($class_roomData['class_room']==null){
+            $class_rooms[] ="undecided";
+        }
+          else{
+        $class_rooms[] = $class_roomData['class_room'] ;} // Access the nested ID
+    }
+    foreach ($data['day'] as $dayData) {
+        $days[] = $dayData['day']; // Access the nested ID
+    }
+    foreach ($data['start_time'] as $start_timeData) {
+        $start_times[] = $start_timeData['start_time']; // Access the nested ID
+    } 
+    for($x=0 ;$x<$numberofentries ;$x++){  //loop should start form zero
+          
+        $scheduleEntry =new ScheduleEntry();
+
+        $scheduleEntry->schedule_id =$schedule->id;
+        $scheduleEntry->instructor_id = $instructorIds[$x];
+        $scheduleEntry->subject_id = $subjectIds[$x];
+        $scheduleEntry->class_room = $class_rooms[$x];
+        $scheduleEntry->day = $days[$x];
+        $scheduleEntry->start_time = $start_times[$x];
+        
+        $scheduleEntry->save();
+    
+    }
+       
+   
+
+
+
+    }
+ 
+    public function update(Request $request ,Schedule $schedule){
+         
+
+        return  $request;
+         
+
+           
+     
+           
+    
+
+        $scheduleEntries = ScheduleEntry::where('schedule_id',$schedule->id)->get();    
+          
+        
+        //return $scheduleEntries[];
+        
+       
+        $data = new Collection($request); 
+       
+        $schedule->name= $data["name"];
+        $schedule->update();
+        $numberofold=0;
+        foreach($scheduleEntries as $entry){
+      
+
+            $dataforupdate =  $data[$entry->id];
+             
+            $entry->instructor_id = $dataforupdate["instructor"];
+            $entry->day = $dataforupdate["day"];
+            $entry->class_room = $dataforupdate["class_room"];
+            $entry->subject_id = $dataforupdate["subject_id"];
+            $entry->start_time = $dataforupdate["start_time"];
+
+            $entry->update();
+
+           $numberofold++; 
+
+
+        } 
+        $numberofnew = $data->count();
+       
+        $numberofnew= $numberofnew-$numberofold;
+        $numberofnew = $numberofnew-2;
+        if($numberofnew ==0)
+        return redirect()->route("schedules");
+    else
+      {
+       
+  
+     
+     $this->add_new($request,$schedule);
+     return redirect()->route('schedules');
+      
+        
+
+}        
+
+       
+
         
     }
     
@@ -198,26 +302,36 @@ class ScheduleEntryController extends Controller
         $schedule= Schedule::find($id);
         if($schedule->delete())
         return redirect()->route('schedules'); 
-        
+   
 
         
      }
      public function delete_entry($id)  {
    
         $scheduleEntry= ScheduleEntry::find($id);
+       
+        if($scheduleEntry != null){
+
+            if($scheduleEntry->delete())
+            //return redirect()->route('edit_schedule', $schedule); 
+            return redirect()->back(); 
+
+        }
+        else{
+
+            return redirect()->back(); 
+
+        }
+          
 
         
-         
-        $scheduleEntry->delete();
-
         
-        ;
-        if($scheduleEntry->delete())
-        return redirect()->back(); 
+      
         
 
         
      }
+ 
 
 
 }
